Order Route and Admin Sales Data API
I have added the Order Route at /api/order/placeOrder. It will create an order in order table, and delete cart. It also works with Authentication, but that is currently uncommented until we connect front and back end.
Also added backend API for sales data, it currently grabs all sales, and puts it into a list of JSON objects that are sorted from best selling items to least sold items. This can be passed to a table for the Admin view.