Fixed local host error + Updates
Updated the home controller with the proper dynamic local host.
Updated login user for when they may have added things to a cart but forgot to log in. If user logs in and they already have a cart in the DB, persist that one. If they do not have a cart, then take the cart they created not logged in and persist that.